fix(tests): cleanup zustand store and split authStore tests
- Adds afterEach hooks to reset store state and clear localStorage to prevent memory leaks between tests.
- Splits the large authStore.test.tsx into smaller, more focused files for login, register, logout, and persistence.
- Fixes typing issues in the test helper .

This addresses the primary cause of the 'JS heap out of memory' error in the test suite.
